requirements.txt
Flask==2.2.5
requests==2.31.0
gunicorn==20.1.0
google-cloud-secret-manager
google-cloud-firestore
